requirements.txt
appdirs==1.4.4
arabic-reshaper==2.1.1
asgiref==3.3.1
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
cryptography==3.4.6
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-appconf==1.0.4
django-cpf==0.1.0
django-imagekit==4.0.2
django-rest-swagger==2.2.0
djangorestframework==3.12.2
easy-thumbnails==2.7.1
future==0.18.2
html5lib==1.1
idna==2.10
isodate==0.6.0
itypes==1.2.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
model-mommy==2.0.0
oauthlib==3.1.0
openapi-codec==1.3.2
pilkit==2.0
Pillow==8.1.2
psycopg2==2.8.6
pycep-correios==5.0.0
pycparser==2.20
PyJWT==2.0.1
PyPDF2==1.26.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2021.1
reportlab==3.5.65
requests==2.25.1
requests-file==1.5.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
xhtml2pdf==0.2.5
zeep==4.0.0
